move namefinding code into its own file
